requirements/base.txt

django==3.0.5
django-environ==0.4.5
django-model-utils==4.0.0
django-allauth==0.41.0
django-crispy-forms==1.9.0 
django-braces==1.14.0
django-model-utils==4.0.0
psycopg2-binary
raven
django-autocomplete-light==3.5.1
django-notifications-hq==1.6.0


